+++ test/bitcrowd_ecto/date_time_test.exs
@@ -1,4 +1,4 @@
 defmodule BitcrowdEcto.DateTimeTest do
   use ExUnit.Case, async: true
-  doctest BitcrowdEcto.DateTime
+  doctest BitcrowdEcto.DateTime, import: true
 end